requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.6
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-recaptcha==3.0.0
idna==3.3
Pillow==8.3.1
pydot==1.4.2
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
